Separeted connection to commerce tools
- Separated commerce tools connection logic
