+++ .gitignore
@@ -16,6 +16,11 @@
 # production
 /build
 
+#env
+.env
+.env.local
+.env.production
+
 # misc
 .DS_Store
 *.pem
